@@ -1,30 +1,23 @@
-buildscript {
-    repositories {
-        maven { url = 'https://maven.minecraftforge.net' }
-        mavenCentral()
-    }
-
-    dependencies {
-        classpath group: 'net.minecraftforge.gradle', name: 'ForgeGradle', version: '5.1.+', changing: true
-    }
+plugins {
+    id 'net.minecraftforge.gradle' version '5.1.+'
 }
 
-apply plugin: 'net.minecraftforge.gradle'
-
-version = '3.0.1'
-group = 'xyz.sirblobman.mod.blobmanstuff' // http://maven.apache.org/guides/mini/guide-naming-conventions.html
+version = '1.19.4-4.0.0'
+group = 'xyz.sirblobman.mod.blobmanstuff'
 archivesBaseName = 'blobmanstuff'
 
-java.toolchain.languageVersion = JavaLanguageVersion.of(8)
+java.toolchain.languageVersion = JavaLanguageVersion.of(17)
 
 minecraft {
-    mappings channel: 'official', version: '1.16.5'
+    mappings channel: 'official', version: '1.19.4'
 
     runs {
         client {
-            workingDirectory project.file('run/client')
+            workingDirectory project.file('run/client/1.19.4')
+
             property 'forge.logging.markers', 'REGISTRIES'
             property 'forge.logging.console.level', 'debug'
+            property 'forge.enabledGameTestNamespaces', 'blobmanstuff'
 
             mods {
                 blobmanstuff {
