@@ -1,21 +1,21 @@
 package xyz.sirblobman.mod.blobmanstuff.entity.ai.goal;
 
 import java.util.EnumSet;
-import java.util.Random;
 
-import net.minecraft.entity.LivingEntity;
-import net.minecraft.entity.ai.controller.MovementController;
-import net.minecraft.entity.ai.goal.Goal;
-import net.minecraft.potion.Effects;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.ai.control.MoveControl;
+import net.minecraft.world.entity.ai.goal.Goal;
 
-import xyz.sirblobman.mod.blobmanstuff.entity.EntityCreeperSlime;
+import xyz.sirblobman.mod.blobmanstuff.entity.CreeperSlime;
 
 public final class SlimeCreeperFaceRandomGoal extends Goal {
-    private final EntityCreeperSlime slime;
+    private final CreeperSlime slime;
     private float chosenDegrees;
     private int nextRandomizeTime;
 
-    public SlimeCreeperFaceRandomGoal(EntityCreeperSlime slime) {
+    public SlimeCreeperFaceRandomGoal(CreeperSlime slime) {
         this.slime = slime;
         setFlags(EnumSet.of(Flag.LOOK));
     }
@@ -27,24 +27,23 @@ public final class SlimeCreeperFaceRandomGoal extends Goal {
         boolean grounded = this.slime.isOnGround();
         boolean water = this.slime.isInWater();
         boolean lava = this.slime.isInWater();
-        boolean levitation = this.slime.hasEffect(Effects.LEVITATION);
+        boolean levitation = this.slime.hasEffect(MobEffects.LEVITATION);
         boolean validLocation = (grounded || water || lava || levitation);
 
-        MovementController moveControl = this.slime.getMoveControl();
+        MoveControl moveControl = this.slime.getMoveControl();
         boolean validControl = (moveControl instanceof SlimeCreeperMovementController);
         return (validTarget && validLocation && validControl);
     }
 
     public void tick() {
         if (--this.nextRandomizeTime <= 0) {
-            Random random = this.slime.getRandom();
+            RandomSource random = this.slime.getRandom();
             this.nextRandomizeTime = 40 + random.nextInt(60);
             this.chosenDegrees = random.nextInt(360);
         }
 
-        MovementController moveControl = this.slime.getMoveControl();
-        if (moveControl instanceof SlimeCreeperMovementController) {
-            SlimeCreeperMovementController controller = (SlimeCreeperMovementController) moveControl;
+        MoveControl moveControl = this.slime.getMoveControl();
+        if (moveControl instanceof SlimeCreeperMovementController controller) {
             controller.setDirection(this.chosenDegrees, false);
         }
     }
