@@ -1,23 +1,23 @@
 package xyz.sirblobman.mod.blobmanstuff.entity.ai.goal;
 
 import java.util.EnumSet;
-import java.util.Random;
 
-import net.minecraft.entity.ai.controller.JumpController;
-import net.minecraft.entity.ai.controller.MovementController;
-import net.minecraft.entity.ai.goal.Goal;
-import net.minecraft.pathfinding.PathNavigator;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.entity.ai.control.JumpControl;
+import net.minecraft.world.entity.ai.control.MoveControl;
+import net.minecraft.world.entity.ai.goal.Goal;
+import net.minecraft.world.entity.ai.navigation.PathNavigation;
 
-import xyz.sirblobman.mod.blobmanstuff.entity.EntityCreeperSlime;
+import xyz.sirblobman.mod.blobmanstuff.entity.CreeperSlime;
 
 public final class SlimeCreeperFloatGoal extends Goal {
-    private final EntityCreeperSlime slime;
+    private final CreeperSlime slime;
 
-    public SlimeCreeperFloatGoal(EntityCreeperSlime slime) {
+    public SlimeCreeperFloatGoal(CreeperSlime slime) {
         this.slime = slime;
         setFlags(EnumSet.of(Flag.JUMP, Flag.MOVE));
 
-        PathNavigator navigation = slime.getNavigation();
+        PathNavigation navigation = slime.getNavigation();
         navigation.setCanFloat(true);
     }
 
@@ -26,7 +26,7 @@ public final class SlimeCreeperFloatGoal extends Goal {
         boolean lava = this.slime.isInLava();
         boolean liquid = (water || lava);
 
-        MovementController moveControl = this.slime.getMoveControl();
+        MoveControl moveControl = this.slime.getMoveControl();
         boolean validControl = (moveControl instanceof SlimeCreeperMovementController);
 
         return (liquid && validControl);
@@ -34,15 +34,14 @@ public final class SlimeCreeperFloatGoal extends Goal {
 
     @Override
     public void tick() {
-        Random random = this.slime.getRandom();
+        RandomSource random = this.slime.getRandom();
         if (random.nextFloat() < 0.8F) {
-            JumpController jumpControl = this.slime.getJumpControl();
+            JumpControl jumpControl = this.slime.getJumpControl();
             jumpControl.jump();
         }
 
-        MovementController moveControl = this.slime.getMoveControl();
-        if (moveControl instanceof SlimeCreeperMovementController) {
-            SlimeCreeperMovementController controller = (SlimeCreeperMovementController) moveControl;
+        MoveControl moveControl = this.slime.getMoveControl();
+        if (moveControl instanceof SlimeCreeperMovementController controller) {
             controller.setWantedMovement(1.2D);
         }
     }
