@@ -1,7 +1,7 @@
 package xyz.sirblobman.mod.blobmanstuff.item;
 
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
 
 import xyz.sirblobman.mod.blobmanstuff.type.SlimeballColor;
 
@@ -12,7 +12,7 @@ public final class ItemColoredSlimeball extends Item {
     private final SlimeballColor color;
 
     public ItemColoredSlimeball(SlimeballColor color) {
-        super(new Properties().tab(BSItemGroups.MAIN).rarity(color.getRarity()));
+        super(new Properties().rarity(color.getRarity()));
         this.color = Validate.notNull(color, "color must not be null!");
     }
 
@@ -23,9 +23,8 @@ public final class ItemColoredSlimeball extends Item {
     @Override
     public boolean isFoil(@NotNull ItemStack stack) {
         Item item = stack.getItem();
-        if (item instanceof ItemColoredSlimeball) {
-            ItemColoredSlimeball itemSlimeball = (ItemColoredSlimeball) item;
-            SlimeballColor color = itemSlimeball.getColor();
+        if (item instanceof ItemColoredSlimeball slimeball) {
+            SlimeballColor color = slimeball.getColor();
             return color.isShiny();
         }
 
