@@ -1,34 +1,33 @@
 package xyz.sirblobman.mod.blobmanstuff.item;
 
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.inventory.EquipmentSlotType;
-import net.minecraft.item.ArmorItem;
-import net.minecraft.item.IArmorMaterial;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.potion.EffectInstance;
-import net.minecraft.potion.Effects;
-import net.minecraft.world.World;
+import net.minecraft.world.effect.MobEffectInstance;
+import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ArmorItem;
+import net.minecraft.world.item.ArmorMaterial;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
 
 import xyz.sirblobman.mod.blobmanstuff.item.material.BlueSlimeArmorMaterial;
 
 public final class ItemBlueSlimeArmor extends ArmorItem {
-    private static final IArmorMaterial BLUE_SLIME = new BlueSlimeArmorMaterial();
+    private static final ArmorMaterial BLUE_SLIME = new BlueSlimeArmorMaterial();
 
-    public ItemBlueSlimeArmor(EquipmentSlotType slot) {
-        super(BLUE_SLIME, slot, new Properties().tab(BSItemGroups.MAIN));
+    public ItemBlueSlimeArmor(Type slot) {
+        super(BLUE_SLIME, slot, new Properties());
     }
 
     @Override
-    public void onArmorTick(ItemStack stack, World world, PlayerEntity player) {
+    public void onArmorTick(ItemStack stack, Level level, Player player) {
         if (isWearingFullArmor(player)) {
-            EffectInstance jumpBoost = new EffectInstance(Effects.JUMP, 20, 1, true,
+            MobEffectInstance jumpBoost = new MobEffectInstance(MobEffects.JUMP, 20, 1, true,
                     false, false);
             player.addEffect(jumpBoost);
         }
     }
 
-    private boolean isWearingFullArmor(PlayerEntity player) {
+    private boolean isWearingFullArmor(Player player) {
         Iterable<ItemStack> armorSlots = player.getArmorSlots();
         for (ItemStack armorSlot : armorSlots) {
             if (armorSlot.isEmpty()) {
