@@ -1,20 +1,19 @@
 package xyz.sirblobman.mod.blobmanstuff.item;
 
-import net.minecraft.item.IItemTier;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Rarity;
-import net.minecraft.item.SwordItem;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Rarity;
+import net.minecraft.world.item.SwordItem;
+import net.minecraft.world.item.Tier;
 
 import xyz.sirblobman.mod.blobmanstuff.item.material.ShinySwordTier;
 
 import org.jetbrains.annotations.NotNull;
 
 public final class ItemShinySword extends SwordItem {
-    private static final IItemTier SHINY_SWORD_TIER = new ShinySwordTier();
+    private static final Tier SHINY_SWORD_TIER = new ShinySwordTier();
 
     public ItemShinySword() {
-        super(SHINY_SWORD_TIER, 3, 0.0F,
-                new Properties().tab(BSItemGroups.MAIN).rarity(Rarity.EPIC));
+        super(SHINY_SWORD_TIER, 3, 0.0F, new Properties().rarity(Rarity.EPIC));
     }
 
     @Override
