@@ -1,9 +1,9 @@
 package xyz.sirblobman.mod.blobmanstuff.item;
 
-import net.minecraft.item.Item;
+import net.minecraft.world.item.Item;
 
 public final class ItemSimple extends Item {
     public ItemSimple() {
-        super(new Properties().tab(BSItemGroups.MAIN));
+        super(new Properties());
     }
 }
