@@ -1,18 +1,18 @@
 package xyz.sirblobman.mod.blobmanstuff.item;
 
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.Items;
-import net.minecraft.item.Rarity;
+import net.minecraft.sounds.SoundEvents;
+import net.minecraft.sounds.SoundSource;
 import net.minecraft.stats.Stats;
-import net.minecraft.util.ActionResult;
-import net.minecraft.util.Hand;
-import net.minecraft.util.SoundCategory;
-import net.minecraft.util.SoundEvents;
-import net.minecraft.world.World;
+import net.minecraft.world.InteractionHand;
+import net.minecraft.world.InteractionResultHolder;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.Rarity;
+import net.minecraft.world.level.Level;
 
-import xyz.sirblobman.mod.blobmanstuff.entity.EntityCustomItemProjectile;
+import xyz.sirblobman.mod.blobmanstuff.entity.ItemProjectile;
 
 import org.jetbrains.annotations.NotNull;
 
@@ -20,8 +20,7 @@ public final class ItemSlimeWand extends Item {
     private final int level;
 
     public ItemSlimeWand(int level) {
-        super(new Properties().stacksTo(1).tab(BSItemGroups.MAIN)
-                .rarity(level == 0 ? Rarity.UNCOMMON : Rarity.RARE));
+        super(new Properties().stacksTo(1).rarity(level == 0 ? Rarity.UNCOMMON : Rarity.RARE));
         this.level = level;
     }
 
@@ -30,33 +29,33 @@ public final class ItemSlimeWand extends Item {
     }
 
     @Override
-    public @NotNull ActionResult<ItemStack> use(@NotNull World world, @NotNull PlayerEntity player,
-                                                @NotNull Hand hand) {
+    public @NotNull InteractionResultHolder<ItemStack> use(@NotNull Level world, @NotNull Player player,
+                                                           @NotNull InteractionHand hand) {
         ItemStack stack = player.getItemInHand(hand);
         double x = player.getX();
         double y = player.getEyeY();
         double z = player.getZ();
 
-        world.playLocalSound(x, y, z, SoundEvents.SLIME_BLOCK_BREAK, SoundCategory.PLAYERS,
+        world.playLocalSound(x, y, z, SoundEvents.SLIME_BLOCK_BREAK, SoundSource.PLAYERS,
                 1.0F, 1.0F, false);
         launchSlime(world, player);
 
         player.awardStat(Stats.ITEM_USED.get(this));
-        return ActionResult.sidedSuccess(stack, world.isClientSide());
+        return InteractionResultHolder.sidedSuccess(stack, world.isClientSide());
     }
 
-    private void launchSlime(World world, PlayerEntity player) {
+    private void launchSlime(Level world, Player player) {
         if (world.isClientSide()) {
             return;
         }
 
         int level = getLevel();
-        Item thrown = (level == 0 ? Items.SLIME_BALL : BSItems.BLUE_SLIME_BALL);
+        Item thrown = (level == 0 ? Items.SLIME_BALL : BSItems.BLUE_SLIME_BALL.get());
         float damage = (2.0F * (level + 1.0F));
 
-        EntityCustomItemProjectile projectile = new EntityCustomItemProjectile(damage, world, player);
-        projectile.shootFromRotation(player, player.xRot, player.yRot,
-                0.0F, 1.5F, 0.0F);
+        ItemProjectile projectile = new ItemProjectile(damage, world, player);
+        projectile.shootFromRotation(player, player.getXRot(), player.getYRot(), 0.0F, 1.5F,
+                0.0F);
 
         ItemStack stack = new ItemStack(thrown, 1);
         projectile.setItem(stack);
