@@ -1,12 +1,13 @@
 package xyz.sirblobman.mod.blobmanstuff.item.material;
 
-import net.minecraft.item.IItemTier;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.crafting.Ingredient;
+import net.minecraft.world.item.Tier;
+import net.minecraft.world.item.crafting.Ingredient;
 
 import xyz.sirblobman.mod.blobmanstuff.block.BSBlocks;
 
-public final class UltraSwordTier implements IItemTier {
+import org.jetbrains.annotations.NotNull;
+
+public final class UltraSwordTier implements Tier {
     @Override
     public int getUses() {
         return (Integer.MAX_VALUE - 1);
@@ -33,8 +34,7 @@ public final class UltraSwordTier implements IItemTier {
     }
 
     @Override
-    public Ingredient getRepairIngredient() {
-        ItemStack stack = new ItemStack(BSBlocks.BLACK_SLIME_BLOCK, 1);
-        return Ingredient.of(stack);
+    public @NotNull Ingredient getRepairIngredient() {
+        return Ingredient.of(BSBlocks.BLACK_SLIME_BLOCK.get());
     }
 }
